Move isRequired to shared-core
